// components/types.ts
export interface IRegister{
    email: string;
    password: string;
    confirmPassword: string;
    firstName: string;
    surname: string;
    buisnessAddress:string;
    state:string;
    mobileNumber:string;
    WNumber:string;
    accountType:string;
    notes:string;
}

export interface IDPEmployee{
    email: string;
    password: string;
    confirmPassword: string;
    firstName: string;
    surname: string;
    
}